Create and test options for filtering stacks
The API for finding leaks accepts a list of functional options to filter
the stacks which should be considered as a leak.

Since background goroutines might not be complete yet, we'll need to
retry and the options have (currently private) ways to control the
retries including number of attempts and duration.