Fix isTraceStack on newer Go versions (#68)
This fixes #66 .

On newer Go versions (1.16 onwards), the top of trace stacks is no longer runtime.goparkunlock. We had a logic that relied on this being the case, which made the tests to fail with -trace flag for newer Go versions. However, to maintain backwards compatibility, we shouldn't just remove this check.

This separates out the isTraceStack into a separate file that gets conditionally compiled on different Go versions (tracestack_old.go for Go 1.15 and before and tracestack_new.go for Go 1.16 and newer) to address this issue.

Internal Ref: GO-887.