Provide stub so apps using goleak can build unchanged with tinygo.
Tinygo can build a large subset of go programs already.
Unfortunately, tinygo currently does not quite support goleak.

As a temporary measure, use the tinygo build tag to provide a stub
for goleak.  This will expand the world of go programs that
run properly on tinygo, and remove one more obstacle keeping
embedded system and wasm users from adopting go and goleak.

This can be reverted if/when tinygo is enhanced to fully
support goleak.