lint: replace deprecated golang.org/x/lint/golint with golangci-lint
golint is deprecated. golangci-lint is a more than adequate
replacement. Because it combines different linters, a single invocation in the
Makefile is enough and then produces consistent output for all of them.

It also gets enabled as a GitHub action. This will annotate pull requests.

The godox linter gets enabled as replacement for the manual "grep -i fixme".

For the sake of convenience, "make lint" installs golangci-lint. In contrast to
golint before, this is done without polluting the go.mod file with the extra
dependencies.