stack: Parse all functions
Adds support to the stack parser
for reading the full list of functions
for a stack trace.

This includes the function that created the stack trace;
it's the bottom of the stack.

We don't maintain the order of the functions
since that's not something we need at this time.
The functions are all placed in a set.