Fix lint and upgrade versions to 1.21/1.22 (#121)
Follow-up to #120, I accidentally merged the PR without waiting for all
tests to run, and missed that 1.20 tests are failing.

Lint is failing because of how we skip golangci-lint run in the action
and use the Makefile. Use the same logic as zap to skip lint.

Tests are failing on 1.20 as stack elision doesn't kick in for the test.
From 1.21 onwards, it does kick in. Rather than try to make the test
trigger stack elision on 1.20, let's bump the versions to 1.21 / 1.22.